Extract `useWorkingDocument` and `useCursorPath` hooks from `QueryBuilder`
